Refactor fiftyone/server/files.py into a package
Split the single-file module into files/{errors,paths,operations}.py
for better separation of concerns. Public API is re-exported from
__init__.py so all existing imports remain unchanged.
